Remove previous networking classes.
