--- Sources/PaymentsCore/Networking/Models/Responses/AccessTokenResponse.swift
@@ -1,7 +1,0 @@
-public struct AccessTokenResponse: Codable {
-    let scope: String
-    let accessToken: String
-    let tokenType: String
-    let expiresIn: Int
-    let nonce: String
-}
