+++ Sources/PaymentsCore/Networking/Protocols/URLSession+URLSessionProtocol.swift
@@ -1,0 +1,9 @@
+
+import Foundation
+
+extension URLSession: URLSessionProtocol {
+    func performRequest(with urlRequest: URLRequest, completionHandler: @escaping (Data?, URLResponse?, Error?) -> Void) {
+        let task = dataTask(with: urlRequest, completionHandler: completionHandler)
+        task.resume()
+    }
+}
