+++ Sources/PaymentsCore/Networking/Protocols/URLSessionProtocol.swift
@@ -1,0 +1,6 @@
+
+import Foundation
+
+protocol URLSessionProtocol {
+    func performRequest(with urlRequest: URLRequest, completionHandler: @escaping (Data?, URLResponse?, Error?) -> Void)
+}
